+++ src/App.tsx
@@ -1,17 +1,10 @@
 import './App.css';
-import { Routes, Route } from 'react-router-dom';
-import { ProtectData, GrantAccess, Navigate } from './features';
+import Front from './features/Front';
 
 function App() {
   return (
     <div id="App">
-      <Routes>
-        <Route path="/" element={<Navigate />}>
-          <Route path="/protectData" element={<ProtectData />} />
-          <Route path="/grantAccess" element={<GrantAccess />} />
-        </Route>
-        <Route path="*" element={<ProtectData />} />
-      </Routes>
+      <Front />
     </div>
   );
 }
