--- src/app/store.ts
@@ -1,23 +1,0 @@
-import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import appReducer from './appSlice';
-import dataProtectedReducer from './dataProtectedSlice';
-
-export const store = configureStore({
-  reducer: {
-    app: appReducer,
-    dataProtected: dataProtectedReducer,
-  },
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware({
-      serializableCheck: false,
-    }),
-});
-
-export type AppDispatch = typeof store.dispatch;
-export type RootState = ReturnType<typeof store.getState>;
-export type AppThunk<ReturnType = void> = ThunkAction<
-  ReturnType,
-  RootState,
-  unknown,
-  Action<string>
->;
