--- src/features/index.ts
@@ -1,3 +1,0 @@
-export { default as ProtectData } from './protectData/ProtectData';
-export { default as GrantAccess } from './grantAccess/GrantAccess';
-export { default as Navigate } from './navigate/Navigate';
