+++ src/main.tsx
@@ -1,28 +1,26 @@
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
-import { BrowserRouter } from "react-router-dom";
-import App from "./App";
-import { ThemeProvider, createTheme } from "@mui/material";
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
+import App from './App';
+import { ThemeProvider, createTheme } from '@mui/material';
 import {
   EthereumClient,
   w3mProvider,
   w3mConnectors,
-} from "@web3modal/ethereum";
-import { WagmiConfig, createClient, configureChains } from "wagmi";
-import { Web3Modal } from "@web3modal/react";
-import { bellecour } from "./utils/walletConnection";
-import { store } from "./app/store";
-import { Provider } from "react-redux";
+} from '@web3modal/ethereum';
+import { WagmiConfig, createClient, configureChains } from 'wagmi';
+import { Web3Modal } from '@web3modal/react';
+import { bellecour } from './utils/walletConnection';
 
-const rootElement = document.getElementById("root");
+const rootElement = document.getElementById('root');
 const root = createRoot(rootElement!);
 
 // material ui theme
 const theme = createTheme({
   palette: {
     primary: {
-      main: "#FCD15A",
-      contrastText: "#1D1D24",
+      main: '#FCD15A',
+      contrastText: '#1D1D24',
     },
   },
 });
@@ -30,7 +28,7 @@ const theme = createTheme({
 // Wagmi Client
 if (!import.meta.env.VITE_WALLET_CONNECT_PROJECT_ID) {
   throw new Error(
-    "You need to provide a WALLET_CONNECT_PROJECT_ID env variable"
+    'You need to provide a WALLET_CONNECT_PROJECT_ID env variable'
   );
 }
 const projectId = import.meta.env.VITE_WALLET_CONNECT_PROJECT_ID!;
@@ -49,14 +47,12 @@ const ethereumClient = new EthereumClient(wagmiClient, chains);
 root.render(
   <StrictMode>
     <ThemeProvider theme={theme}>
-      <Provider store={store}>
-        <BrowserRouter>
-          <WagmiConfig client={wagmiClient}>
-            <App />
-          </WagmiConfig>
-          <Web3Modal projectId={projectId} ethereumClient={ethereumClient} />
-        </BrowserRouter>
-      </Provider>
+      <BrowserRouter>
+        <WagmiConfig client={wagmiClient}>
+          <App />
+        </WagmiConfig>
+        <Web3Modal projectId={projectId} ethereumClient={ethereumClient} />
+      </BrowserRouter>
     </ThemeProvider>
   </StrictMode>
 );
